centralize axios configuration for different requests
