+++ pranama-ui/src/components/View/services/registration_service/RegistrationForm.vue
@@ -130,17 +130,8 @@
 <script>
 import Header from "@/components/View/common/Header";
 import Sidebar from "@/components/View/common/Sidebar";
-import axios from "axios";
-import {API} from "@/shared/index.js"
-const apiClient = axios.create({
-  baseURL: `${API}`,
- 
-  withCredentials: false,
-  headers: {
-    Accept: "application/json",
-    "content-Type": "application/json"
-  }
-});
+import UserService from "@/services/api/UserService";
+
 export default {
   props: {
     msg: String
@@ -192,10 +183,8 @@ export default {
   },
   methods: {
     submitRegistrationForm() {
-      
-      apiClient
-      
-            .post("/api/user/register",{
+      UserService
+        .register({
           email: this.email,
           phone: this.phoneNumber,
           userName: this.userName,
@@ -228,11 +217,11 @@ export default {
    },
 
     showdata() {
-       apiClient
-         .get("/api/user/register")
-         .then(response => {
-     this.regDatas = response.data;
-      console.log(response);
+       UserService
+        .getRegistration()
+        .then(response => {
+          this.regDatas = response.data;
+          console.log(response);
           console.log(this.regDatas);
         })
         .catch(error => {
