+++ pranama-ui/src/services/api/ContactService.js
@@ -1,0 +1,7 @@
+import apiClient from "./axios";
+
+export default {
+  submitContact (params) {
+    return apiClient.post("/api/email/contactus", params);
+  }
+};
