+++ pranama-ui/src/services/api/axios.js
@@ -1,0 +1,13 @@
+import axios from "axios";
+import { API } from "@/shared/index.js";
+
+const apiClient = axios.create({
+  baseURL: `${API}`,
+  withCredentials: false,
+  headers: {
+    Accept: "application/json",
+    "Content-Type": "application/json"
+  }
+});
+
+export default apiClient;
