+++ pranama-ui/src/services/malabar/EmailService.js
@@ -1,0 +1,7 @@
+import apiClient from "./axios";
+
+export default {
+  sendEmail (params) {
+    return apiClient.post("/api/email/contactus", params);
+  }
+};
